+++ utility.js
@@ -412,3 +412,7 @@ export const toSpinalCase = str => str.replace(/[A-Z]/g, (m, i) => i === 0
 export const toCamelCase = str => str.replace(/-[a-z\d]/g, m => m.substr(1).toUpperCase());
 
 export const ucFirst = str => str.substr(0, 1).toUpperCase() + str.substr(1);
+
+export async function filterKeys(obj, keys) {
+	return Object.fromEntries(Object.entries(obj).filter(([key]) => keys.includes(key)));
+}